refactor(web-ui): extract flow chat scroll stability helpers
- move viewport reservation types and calculations into a dedicated module
- keep VirtualMessageList compatible with React Fast Refresh
- update focused tests to import the extracted helpers directly
- preserve existing scroll stabilization behavior
